data-platform-observability/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.51.0
anthropic==0.34.2
pydantic==2.9.2
pandas==2.2.3
